Move the uniqueness validation earlier
move earlier so this happens before the task UI appears
